requirements.txt
Django>=1.10.2
django-webpack-loader>=0.3.3
djangorestframework>=3.4.7
Pillow>=3.4.1
pyotp>=2.2.1
qrcode>=5.3
pygments>=2.1.3
django-guardian>=1.4.6